requirements.txt
Flask==2.2.2
gkeepapi==0.14.2
python-dotenv==0.21.0
requests==2.28.1
schedule==1.1.0
tweepy==4.12.0
